Fix Windows build

+++ lib/ui/compositing/scene.cc
@@ -44,7 +44,8 @@ Scene::Scene(std::shared_ptr<flutter::Layer> rootLayer,
   layer_tree_ = std::make_unique<LayerTree>(
       SkISize::Make(viewport_metrics.physical_width,
                     viewport_metrics.physical_height),
-      viewport_metrics.physical_depth, viewport_metrics.device_pixel_ratio);
+      static_cast<float>(viewport_metrics.physical_depth),
+      static_cast<float>(viewport_metrics.device_pixel_ratio));
   layer_tree_->set_root_layer(std::move(rootLayer));
   layer_tree_->set_rasterizer_tracing_threshold(rasterizerTracingThreshold);
   layer_tree_->set_checkerboard_raster_cache_images(
